Move around/improve validation
