+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildImageMojo.java
@@ -92,8 +92,8 @@ public class BuildImageMojo extends JibPluginConfiguration {
         MavenProjectProperties.getForProject(
             getProject(),
             getLog(),
-            getExtraDirectoryPath(),
-            getExtraDirectoryPermissions(),
+            PluginConfigurationProcessor.getExtraDirectoryPath(this),
+            PluginConfigurationProcessor.convertPermissionsList(getExtraDirectoryPermissions()),
             appRoot);
 
     PluginConfigurationProcessor pluginConfigurationProcessor =
