+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenProjectProperties.java
@@ -21,17 +21,14 @@ import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.cloud.tools.jib.event.JibEventType;
 import com.google.cloud.tools.jib.filesystem.AbsoluteUnixPath;
 import com.google.cloud.tools.jib.frontend.JavaLayerConfigurations;
-import com.google.cloud.tools.jib.maven.JibPluginConfiguration.PermissionConfiguration;
 import com.google.cloud.tools.jib.plugins.common.MainClassInferenceException;
 import com.google.cloud.tools.jib.plugins.common.MainClassResolver;
 import com.google.cloud.tools.jib.plugins.common.ProjectProperties;
 import com.google.cloud.tools.jib.plugins.common.TimerEventHandler;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableMap;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.List;
 import java.util.Map;
 import javax.annotation.Nullable;
 import org.apache.maven.model.Plugin;
@@ -68,15 +65,14 @@ public class MavenProjectProperties implements ProjectProperties {
       MavenProject project,
       Log log,
       Path extraDirectory,
-      List<PermissionConfiguration> permissions,
+      Map<AbsoluteUnixPath, FilePermissions> permissions,
       AbsoluteUnixPath appRoot)
       throws MojoExecutionException {
     try {
       return new MavenProjectProperties(
           project,
           makeEventHandlers(log),
-          MavenLayerConfigurations.getForProject(
-              project, extraDirectory, convertPermissionsList(permissions), appRoot));
+          MavenLayerConfigurations.getForProject(project, extraDirectory, permissions, appRoot));
 
     } catch (IOException ex) {
       throw new MojoExecutionException(
@@ -176,24 +172,5 @@ public class MavenProjectProperties implements ProjectProperties {
     } catch (MainClassInferenceException ex) {
       throw new MojoExecutionException(ex.getMessage(), ex);
     }
-  }
-
-  /**
-   * Validates and converts a list of {@link PermissionConfiguration} to an equivalent {@code
-   * AbsoluteUnixPath->FilePermission} map.
-   *
-   * @param inputList the list to convert
-   * @return the resulting map
-   */
-  @VisibleForTesting
-  static Map<AbsoluteUnixPath, FilePermissions> convertPermissionsList(
-      List<PermissionConfiguration> inputList) {
-    ImmutableMap.Builder<AbsoluteUnixPath, FilePermissions> permissionsMap = ImmutableMap.builder();
-    for (PermissionConfiguration permission : inputList) {
-      AbsoluteUnixPath key = AbsoluteUnixPath.get(permission.getFile());
-      FilePermissions value = FilePermissions.fromOctalString(permission.getMode());
-      permissionsMap.put(key, value);
-    }
-    return permissionsMap.build();
   }
 }
